IA/ai-docs-analytics-credit-card/requirements.txt
streamlit
azure-storage-blob
python-dotenv
azure-ai-documentintelligence
azure-core